+++ src/com/massivecraft/massivecore/cmd/massivecore/CmdMassiveCoreUsysMultiverseList.java
@@ -38,7 +38,7 @@ public class CmdMassiveCoreUsysMultiverseList extends MassiveCommand
 	public void perform() throws MassiveException
 	{
 		// Args
-		int pageHumanBased = (Integer) this.readArg(1);
+		int pageHumanBased = this.readArg(1);
 		
 		// Create Lines
 		List<String> lines = new ArrayList<String>();
